+++ quri_algo/algo/estimator/time_evolution/exact_spectrum.py
@@ -16,7 +16,7 @@ from quri_parts.circuit.transpile import CircuitTranspiler
 from quri_parts.core.estimator import Estimate
 from quri_parts.qulacs.simulator import evaluate_state_to_vector
 
-from quri_algo.core.estimator import State
+from quri_algo.algo.estimator import State
 from quri_algo.problem import HamiltonianInput
 
 from .interface import TimeEvolutionExpectationValueEstimator
