+++ quri_algo/algo/estimator/time_evolution/interface.py
@@ -15,16 +15,16 @@ from quri_parts.circuit.transpile import CircuitTranspiler
 from quri_parts.core.estimator import Estimate
 from quri_parts.core.sampling import Sampler, StateSampler
 
-from quri_algo.circuit.hadamard_test import HadamardTestCircuitFactory
-from quri_algo.circuit.time_evolution.interface import (
-    ControlledTimeEvolutionCircuitFactory,
-)
-from quri_algo.core.estimator import (
+from quri_algo.algo.estimator import (
     ExpectationValueEstimator,
     OperatorPowerEstimatorBase,
     StateT,
 )
-from quri_algo.core.estimator.hadamard_test import HadamardTest
+from quri_algo.algo.estimator.hadamard_test import HadamardTest
+from quri_algo.circuit.hadamard_test import HadamardTestCircuitFactory
+from quri_algo.circuit.time_evolution.interface import (
+    ControlledTimeEvolutionCircuitFactory,
+)
 from quri_algo.problem import HamiltonianT
 
 
@@ -101,8 +101,6 @@ class TimeEvolutionPowerEstimator(OperatorPowerEstimatorBase[HamiltonianT, State
     ):
         self.time_evo_estimator = time_evo_estimator
         self.tau = tau
-
-        self.encoded_operator = self.time_evo_estimator.encoded_operator
         self.transpiler = transpiler
 
     def __call__(
