+++ quri_algo/algo/phase_estimation/spe/gaussian/fitting.py
@@ -14,8 +14,8 @@ import numpy as np
 import numpy.typing as npt
 from scipy import optimize
 
-from quri_algo.core.estimator import OperatorPowerEstimatorBase, StateT
-from quri_algo.core.estimator.time_evolution_estimator import (
+from quri_algo.algo.estimator import OperatorPowerEstimatorBase, StateT
+from quri_algo.algo.estimator.time_evolution import (
     TimeEvolutionExpectationValueEstimator,
     TimeEvolutionPowerEstimator,
 )
